Simplify content util modules
Merge and rename some common module methods
